+++ src/app/app-routing.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+
 import { CartListComponent } from './cart/components/cart-list/cart-list.component';
 import { ProductListComponent } from './products/components/product-list/product-list.component';
 
 const routes: Routes = [
-  { path: '', redirectTo: '/products', pathMatch: 'full' },
   { path: 'products', component: ProductListComponent, title: 'Products' },
-  { path: 'cart', component: CartListComponent, title: 'Cart' }
+  { path: 'cart', component: CartListComponent, title: 'Cart' },
+  { path: '', redirectTo: '/products', pathMatch: 'full' },
+  { path: '**', redirectTo: '/products' }
 ];
 
 @NgModule({