+++ src/app/cart/cart.module.ts
@@ -5,21 +5,20 @@ import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { SharedModule } from '../shared/shared.module';
 import { RouterModule } from '@angular/router';
+import { DeleteAllDialogComponent } from './components/delete-all-dialog/delete-all-dialog.component';
 
 
 
 @NgModule({
   declarations: [
     CartListComponent,
-    CartItemComponent
+    CartItemComponent,
+    DeleteAllDialogComponent
   ],
   imports: [
     CommonModule,
     SharedModule,
     RouterModule
-  ],
-  exports: [
-    CartListComponent
   ]
 })
 export class CartModule { }