+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 
 import { ProductModel } from 'src/app/products/models/product-model';
 import { CartService } from 'src/app/cart/services/cart.service';
 import { CartModel } from '../../models/cart-model';
 import { ProductService } from 'src/app/products/services/product.service';
+import { DeleteAllDialogComponent } from '../delete-all-dialog/delete-all-dialog.component';
 
 @Component({
   selector: 'app-cart-list',
@@ -16,20 +18,52 @@ export class CartListComponent implements OnInit {
   constructor(
     public cartService: CartService,
     private productService: ProductService,
-    ) { }
+    private dialog: MatDialog
+  ) { }
 
   ngOnInit(): void {
-    this.cartItems = this.cartService.getCartItems();
+    this.initCartProducts();
+  }
+  
+  initCartProducts(): void {
+    this.cartItems = this.cartService.getProducts();
+  }
+
+  onQuantityIncrease(product: CartModel): void {
+    this.cartService.increaseQuantity(product);
+  }
+
+  onQuantityDecrease(product: CartModel): void {
+    this.cartService.decreaseQuantity(product);
+  }
+
+  onDeleteItem(product: CartModel): void {
+    this.productService.removeFromCart(product);
+
+    this.cartService.removeProduct(product);
+
+    this.initCartProducts();
+  }
+
+  openDialog(): void {
+    let dialogRef = this.dialog.open(DeleteAllDialogComponent, {
+      width: '30%'
+    })
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        this.removeAll();
+      }
+    })
+  }
+
+  removeAll(): void {
+    this.productService.removeAllFromCart();
+
+    this.cartService.removeAllProducts();
+
+    this.initCartProducts();
   }
 
   trackByItems(index: number, item: ProductModel): number { return item.id; }
-
-  onDeleteItem(item: CartModel): void {
-    this.productService.getProducts().find(el => {
-      if(el.id === item.id) {
-        el.isInCart = false;
-      }
-    });
-    this.cartService.deleteItem(item);
-  }
 }