+++ src/app/cart/components/delete-all-dialog/delete-all-dialog.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-delete-all-dialog',
+  templateUrl: './delete-all-dialog.component.html'
+})
+export class DeleteAllDialogComponent {}