+++ src/app/cart/services/cart.service.ts
@@ -8,46 +8,66 @@ import { CartModel } from '../models/cart-model';
 })
 export class CartService {
 
-  cartItems: CartModel[] = [];
+  private cartProducts: CartModel[] = [];
 
-  getCartItems(): CartModel[] {
-    return this.cartItems;
-  }
-
-  addItem(item: ProductModel, quantity = 1): void {
-
-    for(const cartItem of this.cartItems) {
-      if(cartItem.id === item.id) {
-        cartItem.quantity += quantity;
-        return;
-      }
-    }
-
-    this.cartItems.push({...item, quantity});
-  }
-
-  deleteItem(item: CartModel): void {
-    item.isInCart = false;
-    this.cartItems.splice(this.cartItems.indexOf(item), 1);
-  }
-
-  totalCost(): number {
+  get totalCost(): number {
     let total = 0;
 
-    for(const product of this.cartItems) {
+    for(const product of this.cartProducts) {
       total += product.price * product.quantity;
     }
 
     return total;
   }
 
-  totalQuantity(): number {
+  get totalQuantity(): number {
     let total = 0;
 
-    for(const product of this.cartItems) {
+    for(const product of this.cartProducts) {
       total += product.quantity;
     }
 
     return total;
+  }
+
+  get isEmptyCart(): boolean {
+    return !this.cartProducts.length;
+  }
+
+  getProducts(): CartModel[] {
+    return this.cartProducts;
+  }
+
+  addProduct(product: ProductModel, quantity = 1): void {
+    this.cartProducts = [...this.cartProducts, {...product, quantity}];
+  }
+
+  removeProduct(product: CartModel): void {
+    let newCart = this.cartProducts.filter(el => el.id !== product.id);
+    this.cartProducts = [...newCart];
+  }
+
+  removeAllProducts(): void {
+    this.cartProducts = [];
+  }
+
+  increaseQuantity(product: CartModel): void {
+    this.changeQuantity(product, 1);
+  }
+
+  decreaseQuantity(product: CartModel): void {
+    this.changeQuantity(product, -1);
+  }
+
+  private changeQuantity(product: CartModel, delta: number): void {
+    let updatedCart = this.cartProducts.map(el => {
+      if(el.id === product.id) {
+        product.quantity += delta;
+        return product;
+      }
+      return el;
+    });
+
+    this.cartProducts = [...updatedCart];
   }
 }