+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { CartModel } from 'src/app/cart/models/cart-model';
 import { CartService } from 'src/app/cart/services/cart.service';
 
 import { ProductModel } from 'src/app/products/models/product-model';
@@ -13,6 +14,10 @@ export class ProductListComponent implements OnInit {
   isCartEmpthy!: boolean;
   products!: ProductModel[];
 
+  get cartTooltip(): string {
+    return this.isCartEmpthy ? 'Cart is empthy' : '';
+  }
+
   constructor(
     private productService: ProductService,
     private cartService: CartService
@@ -20,11 +25,16 @@ export class ProductListComponent implements OnInit {
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
-    this.isCartEmpthy = !this.cartService.getCartItems().length;
+    this.isCartEmpthy = this.cartService.isEmptyCart;
   }
 
   onAddToCart(product: ProductModel): void {
-    this.cartService.addItem(product);
+    this.cartService.addProduct(product);
     this.isCartEmpthy = false;
+  }
+
+  onDeleteFromCart(product: ProductModel): void {
+    this.cartService.removeProduct(product as CartModel);
+    this.isCartEmpthy = this.cartService.isEmptyCart;
   }
 }