+++ src/app/products/services/product.service.ts
@@ -8,7 +8,21 @@ import { ProductModel } from '../models/product-model';
 })
 export class ProductService {
 
+  products = products;
+
   getProducts(): ProductModel[] {
-    return products;
+    return this.products;
+  }
+
+  removeFromCart(product: ProductModel): void {
+    this.products.find(el => {
+      if(el.id === product.id) {
+        el.isInCart = false;
+      }
+    });
+  }
+
+  removeAllFromCart(): void {
+    this.getProducts().forEach(el => el.isInCart = false);
   }
 }