+++ src/app/shared/shared.module.ts
@@ -3,22 +3,25 @@ import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { HighlightDirective } from './highlight.directive';
+import { MatDialogModule } from '@angular/material/dialog';
 
+import { HighlightDirective } from './directives/highlight.directive';
+import { ClickDirective } from './directives/click.directive';
 
 
 @NgModule({
   declarations: [
-    HighlightDirective
-  ],
-  imports: [
+    HighlightDirective,
+    ClickDirective
   ],
   exports: [
     MatCardModule,
     MatButtonModule,
     MatIconModule,
     MatTooltipModule,
-    HighlightDirective
+    MatDialogModule,
+    HighlightDirective,
+    ClickDirective
   ]
 })
 export class SharedModule { }